requirements.txt
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
gunicorn==20.0.4
psycopg2==2.8.5
python-dotenv==0.14.0
sqlparse==0.3.1
whitenoise==5.2.0
django-bootstrap4==2.2.0
django-filter==2.3.0
openpyxl==3.0.3
python-dateutil==2.8.1
djangorestframework==3.12.2